added patch test, added migration to link apartments to users

+++ db/migrate/20180919181755_create_apartments.rb
@@ -8,7 +8,7 @@ class CreateApartments < ActiveRecord::Migration[5.2]
       t.string :state
       t.text :country
       t.string :manager_name
-      t.integer :manager_phone
+      t.bigint :manager_phone
       t.text :manager_time
 
       t.timestamps
